+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- `artists.create_figure`: kwarg `margin` allows unique margins on all sides
+
 ### Fixed
 - Fixed bug in `wt.artists.interact2D` where constants were not handled properly.